Combine validation code to use same function in pruned and archive mode.
Fix bug in validation
Add header_row struct
remove difficulties
port over code to the header_row


latests changes

remove test

progress

progress

validation fixes

more progress

more changes

changes

more changes

lmdb changes

more changes

more changes

progeress

remap orphan key index

more fixes

First draft

more cmpile

more compile fixes

more fixes

more compile fixes

proto files some changes

some proto changes

some fixes

proto fixes, changed validation trait

+++ base_layer/core/src/consensus/consensus_constants.rs
@@ -189,14 +189,14 @@ impl ConsensusConstants {
         algos1.insert(PowAlgorithm::Blake, PowAlgorithmConstants {
             max_target_time: 7200,
             min_difficulty: 60_000_000.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 120,
         });
         algos1.insert(PowAlgorithm::Monero, PowAlgorithmConstants {
             max_target_time: 7200,
             min_difficulty: 1.into(),
 
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 120,
         });
 
@@ -204,13 +204,13 @@ impl ConsensusConstants {
         algos2.insert(PowAlgorithm::Blake, PowAlgorithmConstants {
             max_target_time: 7200,
             min_difficulty: 1.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 120,
         });
         algos2.insert(PowAlgorithm::Monero, PowAlgorithmConstants {
             max_target_time: 7200,
             min_difficulty: 1.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 120,
         });
 
@@ -218,13 +218,13 @@ impl ConsensusConstants {
         algos3.insert(PowAlgorithm::Blake, PowAlgorithmConstants {
             max_target_time: 7200,
             min_difficulty: 60_000_000.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 120,
         });
         algos3.insert(PowAlgorithm::Monero, PowAlgorithmConstants {
             max_target_time: 7200,
             min_difficulty: 59_000.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 120,
         });
 
@@ -232,13 +232,13 @@ impl ConsensusConstants {
         algos4.insert(PowAlgorithm::Blake, PowAlgorithmConstants {
             max_target_time: 720,
             min_difficulty: 60_000_000.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 120,
         });
         algos4.insert(PowAlgorithm::Monero, PowAlgorithmConstants {
             max_target_time: 720,
             min_difficulty: 59_000.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 120,
         });
         vec![
@@ -349,13 +349,13 @@ impl ConsensusConstants {
         algos.insert(PowAlgorithm::Sha3, PowAlgorithmConstants {
             max_target_time: 1800,
             min_difficulty: 60_000_000.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 300,
         });
         algos.insert(PowAlgorithm::Monero, PowAlgorithmConstants {
             max_target_time: 1200,
             min_difficulty: 60_000.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 200,
         });
         vec![ConsensusConstants {
@@ -382,13 +382,13 @@ impl ConsensusConstants {
         algos.insert(PowAlgorithm::Sha3, PowAlgorithmConstants {
             max_target_time: 1800,
             min_difficulty: 60_000_000.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 300,
         });
         algos.insert(PowAlgorithm::Monero, PowAlgorithmConstants {
             max_target_time: 800,
             min_difficulty: 60_000_000.into(),
-            max_difficulty: u64::MAX.into(),
+            max_difficulty: u128::MAX.into(),
             target_time: 200,
         });
         vec![ConsensusConstants {
